+++ src/platforms/mesa/server/kms/real_kms_display_configuration.cpp
@@ -171,9 +171,9 @@ void mgm::RealKMSDisplayConfiguration::add_or_update_output(
         kms_connector_type_to_output_type(connector.connector_type)};
     geom::Size physical_size{connector.mmWidth, connector.mmHeight};
     bool connected{connector.connection == DRM_MODE_CONNECTED};
-    size_t const invalid_mode_index = std::numeric_limits<size_t>::max();
-    size_t current_mode_index{invalid_mode_index};
-    size_t preferred_mode_index{invalid_mode_index};
+    uint32_t const invalid_mode_index = std::numeric_limits<uint32_t>::max();
+    uint32_t current_mode_index{invalid_mode_index};
+    uint32_t preferred_mode_index{invalid_mode_index};
     std::vector<DisplayConfigurationMode> modes;
     std::vector<MirPixelFormat> formats {mir_pixel_format_argb_8888,
                                          mir_pixel_format_xrgb_8888};
